+++ wazo_calld/plugins/calls/http.py
@@ -287,13 +287,17 @@ class ConnectCallToUserResource(AuthResource):
 
     @required_acl('calld.calls.{call_id}.user.{user_uuid}.update')
     def put(self, call_id, user_uuid):
+        tenant = Tenant.autodetect()
         body = connect_call_request_body_schema.load(
             request.json if request.data else {}
         )
-        assert body
-        args = {'call_id': call_id, 'user_uuid': user_uuid, 'timeout': body['timeout']}
 
-        new_call_id = self.calls_service.connect_user(**args)
+        new_call_id = self.calls_service.connect_user(
+            call_id=call_id,
+            user_uuid=user_uuid,
+            timeout=body['timeout'],
+            tenant_uuid=tenant.uuid,
+        )
         new_call = self.calls_service.get(new_call_id)
 
         return call_schema.dump(new_call)